+++ src/Domain/JsonApi/V1/Server.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Dystcz\LunarApiNewsletter\Domain\JsonApi\V1;
+namespace Dystore\Newsletter\Domain\JsonApi\V1;
 
-use Dystcz\LunarApi\Domain\JsonApi\Servers\Server as BaseServer;
-use Dystcz\LunarApiNewsletter\Domain\Newsletter\JsonApi\V1\NewsletterSchema;
+use Dystore\Api\Domain\JsonApi\Servers\Server as BaseServer;
+use Dystore\Newsletter\Domain\Newsletter\JsonApi\V1\NewsletterSchema;
 
 class Server extends BaseServer
 {