+++ src/Domain/Newsletter/Http/Controllers/SubscribeToNewsletterController.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Dystcz\LunarApiNewsletter\Domain\Newsletter\Http\Controllers;
+namespace Dystore\Newsletter\Domain\Newsletter\Http\Controllers;
 
-use Dystcz\LunarApi\Base\Controller;
-use Dystcz\LunarApiNewsletter\Domain\Newsletter\Actions\SubscribeToNewsletter;
-use Dystcz\LunarApiNewsletter\Domain\Newsletter\JsonApi\V1\NewsletterSubscriptionRequest;
+use Dystore\Api\Base\Controller;
+use Dystore\Newsletter\Domain\Newsletter\Actions\SubscribeToNewsletter;
+use Dystore\Newsletter\Domain\Newsletter\JsonApi\V1\NewsletterSubscriptionRequest;
 use Illuminate\Http\Response;
 
 class SubscribeToNewsletterController extends Controller