+++ src/Domain/Newsletter/JsonApi/V1/NewsletterResource.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Dystcz\LunarApiNewsletter\Domain\Newsletter\JsonApi\V1;
+namespace Dystore\Newsletter\Domain\Newsletter\JsonApi\V1;
 
-use Dystcz\LunarApi\Domain\JsonApi\Resources\JsonApiResource;
+use Dystore\Api\Domain\JsonApi\Resources\JsonApiResource;
 use Illuminate\Http\Request;
 
 class NewsletterResource extends JsonApiResource