+++ src/LunarApiNewsletter.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Dystcz\LunarApiNewsletter;
-
-class LunarApiNewsletter
-{
-    // Build your next great package.
-}