+++ src/Newsletter.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Dystore\Newsletter;
+
+class Newsletter
+{
+    // Build your next great package.
+}