+++ tests/Pest.php
@@ -1,8 +1,0 @@
-<?php
-
-function serverUrl(?string $path = null, bool $full = false): string
-{
-    $path = implode('/', [Config::get('lunar-api.general.route_prefix'), 'v1', ltrim($path, '/')]);
-
-    return $path;
-}