+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
-All notable changes to `lunar-api-newsletter` will be documented in this file
+All notable changes to `dystore-newsletter` will be documented in this file
 
 ## 1.0.0 - 201X-XX-XX
 
-- initial release
+-   initial release